pyproject.toml
# pyproject.toml

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bots-control-panel"
version = "1.0.0"
description = "FastAPI app to control multiple systemd services via HTTP API with unique API keys."
readme = "README.md"
requires-python = ">=3.8"

classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: FastAPI",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Database",
    "Topic :: Internet :: WWW/HTTP :: WSGI :: Application",
    "Topic :: System :: Systems Administration",
]

# --- Core Runtime Dependencies ---
dependencies = [
    "jinja2>=3.0.0,<4.0.0",
    "fastapi>=0.100.0,<1.0.0",
    "uvicorn[standard]>=0.20.0,<1.0.0",
    "sqlalchemy>=2.0.0,<3.0.0",
    "pydantic>=2.0.0,<3.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "psycopg2-binary>=2.9.0,<3.0.0"
]

# --- Optional Dependencies ---
[project.optional-dependencies]
dev = [
    "pytest>=7.0.0,<9.0.0",
    "httpx>=0.24.0,<1.0.0",
    "ruff",
]
